Moved main.py to a Python package
